compend/pyproject.toml
[tool.poetry]
name = "compend"
version = "0.1.0"
description = ""
authors = ["Ez Ablaza <ezra.ablaza.inbox@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
discord-py = "~2.1.0"
requests = "~2.28.1"


[tool.poetry.group.dev.dependencies]
pytest = "~7.2.0"
black = "~22.12.0"
pytest-asyncio = "~0.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
